fuse usage command into pokemon
